Manually fix measurement terms with Equivalent Class Axiom lacking genus #987
Combined genus:measurement with existing is_about clauses for measurements that lacked this as per Nico's spreadsheet. Relaxed terms to subClassOf axiom for measurements mainly for disease biomarkers but I am not that happy with doing this because several children are not part of this superclass after doing so. Commented with details on the ticket.